Added an installation guide and TODO.
The TODO list is a list of tasks that needs to be accomplished before I
make the repository public. This includes mainly documentation, a few
bug fixes and implementation of a few simple features. The installation
guide assumes some things that are yet to have been implemented, but
those tasks should be included in the TODO.